Renamed and moved sync folder & code refactored
