+++ client/src/app/components/sync/sync.component.html
@@ -26,6 +26,7 @@
                     <br />
                 </tr>
             </tbody>
+            <tr>Total Steps: {{ totalStepCount }}</tr>
         </table>
     </div>
     <div *ngIf="!stepCounts.length && isStepsCounted">
