+++ client/src/app/components/sync/sync.component.less
