--- client/src/app/home/home.component.ts
@@ -1,63 +1,0 @@
-import { Component, OnInit, NgZone } from '@angular/core';
-import { GoogleFitService } from '../services/google-fit.service';
-
-@Component({
-	selector: 'app-home',
-	templateUrl: './home.component.html',
-	styleUrls: ['./home.component.less']
-})
-export class HomeComponent implements OnInit {
-	private stepResp = null;
-	private isStepsCounted = false;
-	private stepCounts = [];
-	private playerGmail;
-
-	constructor(
-		private ngZone: NgZone,
-		private googleFitService: GoogleFitService
-	) { }
-
-	ngOnInit() { }
-
-	signIn() {
-		this.googleFitService
-			.init()
-			.then(() => {
-				console.log('inited');
-			})
-			.catch(err => {
-				console.log(`${err} init`);
-			});
-	}
-
-	viewStepCount() {
-		const timeGap = {
-			endTimeMillis: +new Date(),
-			startTimeMillis: 1561228200000
-		};
-		this.googleFitService.checkCount(timeGap).subscribe(
-			resp => {
-				console.log(resp);
-				resp.bucket.forEach(element => {
-					if (element.dataset[0].point[0]) {
-						this.stepCounts.push({
-							dateSteps: new Date(+element.endTimeMillis),
-							stepsPerDay: element.dataset[0].point[0].value[0].intVal
-						});
-					}
-				});
-				this.ngZone.run(() => {
-					this.isStepsCounted = true;
-				});
-			},
-			err => {
-				console.log('error view steps');
-			}
-		);
-	}
-
-	syncWithApi() {
-		this.googleFitService.syncData(this.stepCounts);
-	}
-
-}
