+++ client/src/index.html
@@ -1,5 +1,4 @@
 <!doctype html>
-<script src="https://apis.google.com/js/api.js"></script>
 <html lang="en">
     <head>
         <meta charset="utf-8">
@@ -14,6 +13,8 @@
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
         <meta name="description" content="">
         <meta name="author" content="">
+        <!-- Google API -->
+        <script src="https://apis.google.com/js/api.js"></script>
         <!-- Core CSS -->
         <link href="assets/css/mains.css" rel="stylesheet">
         <!-- Fontello -->
